Add files via upload
This work highlights a comparison between three different situations in a very simple fuel cycle model: 
1.	No traps in the Breeding Blanket
2.	The trap density varies linearly with the logarithm of the radiation damage [dpa]
3.	The trap density varies linearly with the radiation damage [dpa]
The target is to obtain a table with values of Inventory Startup and TBRr in the three different scenarios. 
In this very simple analysis, the presence of traps impacts only the residence time in the Breeding Blanket.  It is initially set equal to 1 hour; it increases during the operational period of the reactor (2 years, corresponding to the doubling time).